refactor: first iteration of moving graphing logic outside of the presentation
